requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

diffq==0.2.4
numpy==1.26.4
pydub==0.25.1
demucs==4.0.1
stempeg==0.2.3
lameenc==1.7.0
torch==2.2.1+cpu
torchaudio==2.2.1+cpu
python-telegram-bot==20.8